Move System.Text.Json tests into dedicated test project.
