+++ Quantities.sln
@@ -17,6 +17,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "quantities.serialization.ne
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "quantitier.serialization.newtonsoft.test", "quantities.serialization\newtonsoft.test\quantitier.serialization.newtonsoft.test.csproj", "{A8F2E897-0E78-4675-8A02-448EC8A61388}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "quantities.serialization.text.json.test", "quantities.serialization\text.json.test\quantities.serialization.text.json.test.csproj", "{1C0B88CF-0A15-4128-9643-55945694C522}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -50,10 +52,15 @@ Global
 		{A8F2E897-0E78-4675-8A02-448EC8A61388}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A8F2E897-0E78-4675-8A02-448EC8A61388}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A8F2E897-0E78-4675-8A02-448EC8A61388}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1C0B88CF-0A15-4128-9643-55945694C522}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1C0B88CF-0A15-4128-9643-55945694C522}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1C0B88CF-0A15-4128-9643-55945694C522}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1C0B88CF-0A15-4128-9643-55945694C522}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{64C99937-7C3C-437A-BB52-CDF80F156EEE} = {E04095D5-F6D9-467F-9D7E-CC7AF8396E9F}
 		{77CD9A1A-1965-414B-B6A4-C580448E46BA} = {E04095D5-F6D9-467F-9D7E-CC7AF8396E9F}
 		{A8F2E897-0E78-4675-8A02-448EC8A61388} = {E04095D5-F6D9-467F-9D7E-CC7AF8396E9F}
+		{1C0B88CF-0A15-4128-9643-55945694C522} = {E04095D5-F6D9-467F-9D7E-CC7AF8396E9F}
 	EndGlobalSection
 EndGlobal
