+++ quantities.serialization/text.json.test/ISerializationTester.cs
@@ -1,0 +1,7 @@
+namespace Quantities.Serialization.Text.Json.Text;
+
+internal interface ISerializationTester<TQuantity>
+{
+    void SupportsSerialization(TQuantity quantity);
+    static abstract IEnumerable<Object[]> Quantities();
+}
