+++ quantities.serialization/text.json.test/Usings.cs
@@ -1,0 +1,6 @@
+global using Quantities.Prefixes;
+global using Quantities.Quantities;
+global using Quantities.Units.Si;
+global using Quantities.Units.Si.Derived;
+global using Quantities.Units.Si.Metric;
+global using Xunit;
