+++ quantities.test/Convenience.cs
@@ -1,38 +1,14 @@
 using System.Globalization;
-using System.Text.Json;
-using Quantities.Measures;
-using Quantities.Serialization.Text.Json;
 using Xunit.Sdk;
 
 namespace Quantities.Test;
 public static class Convenience
 {
-    private static readonly JsonSerializerOptions options = new JsonSerializerOptions().EnableQuantities();
     private const Int32 fullPrecision = 16;
     public static Int32 FullPrecision => fullPrecision;
     public static Int32 MediumPrecision => fullPrecision - 1;
     public static Int32 LowPrecision => fullPrecision - 2;
     public static Int32 VeryLowPrecision => fullPrecision - 3;
-
-    internal static void HasSameMeasure(this Quant quant, in Quant other)
-    {
-        Assert.True(quant.MeasureEquals(other), $"Mismatching measures on '{quant}' and '{other}'");
-    }
-
-    public static T SupportsSerialization<T>(this T value)
-        where T : IEquatable<T>
-    {
-        var serialized = Serialize(value);
-        var deserialized = Deserialize<T>(serialized);
-
-        Assert.Equal(value, deserialized);
-        return deserialized;
-    }
-    public static T SerializeRoundRobin<T>(this T value) => Deserialize<T>(Serialize(value));
-    public static String Serialize<T>(this T value) => Serialize(value, options);
-    public static String Serialize<T>(this T value, JsonSerializerOptions options) => JsonSerializer.Serialize(value, options);
-    public static T Deserialize<T>(String value) => Deserialize<T>(value, options);
-    public static T Deserialize<T>(String value, JsonSerializerOptions options) => JsonSerializer.Deserialize<T>(value, options) ?? throw new Exception($"Deserialization of {typeof(T).Name} failed");
     public static void Matches<TQuantity>(this TQuantity actual, TQuantity expected)
         where TQuantity : struct, IQuantity<TQuantity>, Dimensions.IDimension
     {
