+++ quantities.test/ElectricCurrentTest.cs
@@ -54,18 +54,4 @@ public sealed class ElectricCurrentTest
 
         current.Matches(expected);
     }
-    [Theory]
-    [MemberData(nameof(Currents))]
-    public void ElectricCurrentSupportsSerialization(ElectricCurrent current) => current.SupportsSerialization();
-
-    public static IEnumerable<Object[]> Currents()
-    {
-        static IEnumerable<ElectricCurrent> Interesting()
-        {
-            yield return ElectricCurrent.Of(21).Si<Ampere>();
-            yield return ElectricCurrent.Of(342).Si<Femto, Ampere>();
-            yield return ElectricCurrent.Of(6).Si<Kilo, Ampere>();
-        }
-        return Interesting().Select(l => new Object[] { l });
-    }
 }
