+++ quantities.test/TemperatureTest.cs
@@ -271,23 +271,4 @@ public sealed class TemperatureTest
 
         actual.Matches(expected);
     }
-
-    [Theory]
-    [MemberData(nameof(Temperatures))]
-    public void TemperatureSupportsSerialization(Temperature temperature) => temperature.SupportsSerialization();
-
-    public static IEnumerable<Object[]> Temperatures()
-    {
-        static IEnumerable<Temperature> Interesting()
-        {
-            yield return Temperature.Of(21).Si<Kelvin>();
-            yield return Temperature.Of(342).Si<Nano, Kelvin>();
-            yield return Temperature.Of(6).Metric<Celsius>();
-            yield return Temperature.Of(-41).Metric<Milli, Celsius>();
-            yield return Temperature.Of(1.21).Imperial<Fahrenheit>();
-            yield return Temperature.Of(121).Imperial<GasMark>();
-            yield return Temperature.Of(95.2).NonStandard<Delisle>();
-        }
-        return Interesting().Select(l => new Object[] { l });
-    }
 }
