+++ quantities.test/TimeTest.cs
@@ -82,22 +82,4 @@ public sealed class TimeTest
 
         actual.Matches(expected);
     }
-    [Theory]
-    [MemberData(nameof(Times))]
-    public void TimeSupportsSerialization(Time time) => time.SupportsSerialization();
-
-    public static IEnumerable<Object[]> Times()
-    {
-        static IEnumerable<Time> Interesting()
-        {
-            yield return Time.Of(21).Si<Second>();
-            yield return Time.Of(342).Si<Ronto, Second>();
-            yield return Time.Of(6).Si<Deci, Second>();
-            yield return Time.Of(-41).Metric<Minute>();
-            yield return Time.Of(1.21).Metric<Hour>();
-            yield return Time.Of(121).Metric<Day>();
-            yield return Time.Of(95.2).Metric<Week>();
-        }
-        return Interesting().Select(l => new Object[] { l });
-    }
 }
