+++ quantities.test/VelocityTest.cs
@@ -1,5 +1,4 @@
-﻿using Quantities.Units.NonStandard.Length;
-using Quantities.Units.Si.Metric;
+﻿using Quantities.Units.Si.Metric;
 
 namespace Quantities.Test;
 
@@ -102,24 +101,5 @@ public sealed class VelocityTest
         Velocity actual = length / time;
 
         Assert.Equal(expected, actual);
-    }
-
-    [Theory]
-    [MemberData(nameof(Velocities))]
-    public void VelocitySupportsSerialization(Velocity velocity) => velocity.SupportsSerialization().Quant.HasSameMeasure(velocity.Quant);
-
-    public static IEnumerable<Object[]> Velocities()
-    {
-        static IEnumerable<Velocity> Interesting()
-        {
-            yield return Velocity.Of(21).Si<Metre>().Per.Si<Second>();
-            yield return Velocity.Of(342).Si<Kilo, Metre>().Per.Metric<Hour>();
-            yield return Velocity.Of(6).Imperial<Mile>().Per.Metric<Hour>();
-            yield return Velocity.Of(-41).Imperial<Foot>().Per.Si<Second>();
-            yield return Velocity.Of(1.21).Metric<Ångström>().Per.Si<Micro, Second>();
-            yield return Velocity.Of(0.125).Metric<AstronomicalUnit>().Per.Metric<Minute>();
-            yield return Velocity.Of(0.000292002383).NonStandard<LightYear>().Per.Metric<Week>();
-        }
-        return Interesting().Select(l => new Object[] { l });
     }
 }
