+++ quantities/IMeasureEquals.cs
@@ -1,0 +1,7 @@
+namespace Quantities;
+
+internal interface IMeasureEquals<TSelf>
+    where TSelf : struct, IMeasureEquals<TSelf>
+{
+    Boolean EqualMeasure(in TSelf other);
+}
